thea: camera: disable zsl
To achieve 'zero-shutter-lag' the camera driver has to keep a full
ress copy of the current preview in a buffer. When pictures are
taken in dark environments and flash is required. For normal capt-
ures, it's common to use a double flash to calibrate the sensor on
the first and capture on the second.

After the recent snap changes, captures that trigger flash are
lacking the second flash when zsl is enabled. This leads to the
pictures being very dark.

It may be entirely possible that our sensor is to slow to cali-
brate and shot within 1 flash period or that zsl was never meant
to be used with flash.

Disable zsl entirely while we find out a proper fix.

Change-Id: Ia57102fa332e2f6bdb3e3a8afcf538116248b628